implement ability to pass flags to compiler and bin from test
the top two lines of a test case can be interpreted as flags to pass to
the compile when compiling the test and flags to pass to the executable
when running the test.

// -test [flags...]
the flags part of this comment will be passed to the executable when
running the test.

// -compile [flags...]
the flags part of this comment will be passed to the compiler when
compiling the test.