'self' is now a pseudo-keyword in the context of struct methods
when 'self' is used as the first parameter for a struct method, it
will automatically have its type set to the enclosing struct reference.

this also allows for the automatic 'self' insertion to check if the
type of the first parameter matches the type of the object.

e.g:
struct Foo in
    def bar self in
        ...
    end

    def baz self: int in
        // this is valid
    end
end

`foo.bar()` rewrites into `Foo::bar(foo)` and
`foo.baz(123)` rewrites into `Foo::baz(123)`