infer specific type for non-mono-type varargs when calling mono-type
`
def foo ...ints: int in
    ...
end
def bar ...args in
    foo(args);
end
`
In the above example, when forward non-mono-type varargs to a function
with mono-type varargs we could trick the type system into allowing this
to be valid code. Now, the type inference will specify bar's `args`
param to be a 'tuple of int' mono-type tuple.